% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehmcexamples-package.R
\docType{package}
\name{ehmcexamples-package}
\alias{ehmcexamples-package}
\alias{ehmcexamples}
\title{The 'ehmcexamples' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.7. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://arxiv.org/abs/1810.04449}
}

}
\author{
\strong{Maintainer}: Julien Stoehr \email{stoehr@ceremade.dauphine.fr} (\href{https://orcid.org/0000-0002-7813-0185}{ORCID})

}
