% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stanmodels.R
\name{get_stanmodels}
\alias{get_stanmodels}
\title{Getters of the precompiled stan models}
\usage{
get_stanmodels()
}
\value{
A list of object of class \code{stanfit}
}
\description{
Getters of the precompiled stan models
}
